+++ climpred/tests/test_perfectModelEnsemble_class.py
@@ -49,7 +49,7 @@ def test_compute_persistence(perfectModelEnsemble_initialized_control):
 @pytest.mark.slow
 def test_bootstrap(perfectModelEnsemble_initialized_control):
     """Test that perfect model ensemble object can be bootstrapped"""
-    perfectModelEnsemble_initialized_control.bootstrap(bootstrap=2)
+    perfectModelEnsemble_initialized_control.bootstrap(iterations=2)
 
 
 def test_get_initialized(PM_ds_initialized_1d):